Была переделана архитектура
