--- laboratory-task-16/WORKSPACE
@@ -1,1 +1,0 @@
-
