+++ laboratory-task-16/src/func/BUILD
@@ -1,0 +1,8 @@
+cc_library(
+    name = "func",
+    srcs = ["func.cpp"],
+    hdrs = ["func.hpp"],
+    data = ["//src/files:info"],
+    deps = ["//src/class:class"],
+    visibility = ["//visibility:public"]
+)
