+++ laboratory-task-16/src/main/BUILD
@@ -1,6 +1,6 @@
 cc_binary(
     name = "cpp_app",
     srcs = ["main.cpp"],
-    data = ["//src/files:info"],
-    deps = ["//src/class:class"]
+    
+    deps = ["//src/func:func"]
 )
