+++ laboratory-task-16/src/tests/BUILD
@@ -1,10 +1,8 @@
-## src/tests/BUILD
-
 cc_test(
     name = "tests",
     srcs = ["tests.cpp"],
     deps = [
         "@googletest//:gtest_main",
-        "//src/class:class",
+        "//src/func:func",
     ]
 )
